Make State non-copyable & non-moveable; store states in a forward_list
std::vector does not work with non-copyable and non-moveable States.
I had to write a `State(State&&) { throw; }` c-tor if I wanted to use
std::vector, because a vector must be resizable, while I cannot allow
this with my States.

(copyable or moveable States are dangerous because I have pointers to
 State in State, and errors with this take me hours of searching for
 bugs)