Add a char_type typedef to Alphabet
To avoid explicitly writing out the char type of an alphabet later.
(i.e. instead of writing CT, we'll be able to write alphabet::char_type)
It's useful when some alphabet's type is deduced and not explicitly
written, i.e. when the char type template argument is not clear from the
context.